R2GenCSR/requirements.txt
absl-py==2.1.0
accelerate==0.29.2
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
auto_gptq==0.7.1
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
boto3==1.34.81
botocore==1.34.81
buildtools==1.0.6
captum==0.7.0
causal-conv1d==1.2.0.post2
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm @ file:///croot/comm_1709322850197/work
constantly==23.10.4
contourpy==1.2.1
croniter==1.4.1
cycler==0.12.1
datasets==2.19.1
dateutils==0.6.12
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepspeed==0.14.0
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
easydict==1.13
editor==1.6.6
einops==0.7.0
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.9.0
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
furl==2.1.3
fvcore==0.1.5.post20221221
gekko==1.1.1
grad-cam==1.5.2
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
inquirer==3.2.4
iopath==0.1.10
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython==7.34.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lightning==2.0.5
lightning-cloud==0.5.65
lightning-utilities==0.11.2
line_profiler==4.1.3
mamba-ssm==1.2.0.post1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
optimum==1.19.2
ordered-set==4.1.0
orderedmultidict==1.0.1
orjson==3.10.0
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.1
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.10.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare==0.7.5
pillow==10.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
pluggy==1.4.0
portalocker==2.8.2
prometheus_client==0.20.0
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
protobuf==5.26.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.16.3
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
qtconsole==5.5.1
QtPy==2.4.1
readchar==4.0.6
redo==2.0.4
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.0
runs==1.2.2
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selective-scan @ file:///VMamba/kernels/selective_scan
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.37.2
starsessions==1.3.0
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
timm==0.4.12
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchcam==0.4.0
torchmetrics==1.3.2
torchvision==0.17.0+cu121
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.66.2
traitlets==5.14.2
transformers==4.40.2
transformers-interpret==0.10.0
triton==2.2.0
ttach==0.0.3
Twisted==24.3.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2024.1
uri-template==1.3.0
urllib3==1.26.13
uvicorn==0.29.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
xmod==1.8.1
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.18.1
zope.interface==6.3
